FEAT #13 : KafkaProducer 스트림 데이터 처리 로직 수정
