+++ src/main/java/com/skuniv/fuwarilog/kafka/KafkaProducerService.java
@@ -18,9 +18,12 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestTemplate;
 import org.springframework.web.util.UriComponentsBuilder;
 
+import java.net.URI;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
+import java.util.Arrays;
+import java.util.List;
 
 @Service
 @Slf4j
@@ -36,21 +39,24 @@ public class KafkaProducerService {
     /**
      * @implSpec 실시간 환율 데이터 연동 및 전달
      */
-    @Scheduled(cron = "0 0 11 * * *")
+    @Scheduled(cron = "0 0/30 * * * *")
     public void fetchAndSendExchangeRates() {
         // 환율 실시간 API 연동
-        String url = UriComponentsBuilder.fromHttpUrl("https://www.koreaexim.go.kr/site/program/financial/exchangeJSON")
+        String uri = UriComponentsBuilder.fromHttpUrl("https://www.koreaexim.go.kr/site/program/financial/exchangeJSON")
                 .queryParam("authkey", apiKey)
-                //.queryParam("searchdate", "20250522")
-                .queryParam("searchdate", LocalDate.now().format(DateTimeFormatter.ofPattern("yyyyMMdd")))
+                .queryParam("searchdate", LocalDate.now().format(DateTimeFormatter.ofPattern("yyyy-MM-dd")))
                 .queryParam("data", "AP01")
                 .toUriString();
 
         log.info("Fetching exchange rates from Kafka ...");
-        log.info(url);
+        log.info(uri);
 
         try {
-            ResponseEntity<String> response = restTemplate.getForEntity(url, String.class);
+            ResponseEntity<String> response = restTemplate.getForEntity(uri, String.class);
+
+            log.info(response.getBody());
+
+            List<String> currency = Arrays.asList("USD", "JPY", "CNY", "KRW");
 
             if (response.getStatusCode() == HttpStatus.OK) {
                 ObjectMapper mapper = new ObjectMapper();
@@ -61,14 +67,26 @@ public class KafkaProducerService {
                 for (JsonNode node : root) {
                     if(!node.path("result").asText().equals("1")) continue;
 
-                    ExchangeRateRequest.ExchangeRateDTO dto = new ExchangeRateRequest.ExchangeRateDTO();
-                    dto.setCurUnit(node.get("cur_unit").asText());
-                    dto.setDealBasR(node.get("deal_bas_r").asText());
-                    dto.setCurNm(node.get("cur_nm").asText());
-                    dto.setTimeStamp(LocalDateTime.now().toString());
+                    if(currency.contains(node.path("cur_unit").asText())) {
 
-                    String json = mapper.writeValueAsString(dto);
-                    kafkaTemplate.send("exchange_rate", dto.getCurUnit(), json);
+                        ExchangeRateRequest.ExchangeRateDTO dto = new ExchangeRateRequest.ExchangeRateDTO();
+                        dto.setCurUnit(node.get("cur_unit").asText());
+                        dto.setDealBasR(node.get("deal_bas_r").asText());
+                        dto.setCurNm(node.get("cur_nm").asText());
+                        dto.setTimeStamp(LocalDateTime.now().toString());
+
+                        log.info("Sending data to Kafka topic...");
+
+                        String json = mapper.writeValueAsString(dto);
+                        kafkaTemplate.send("exchange_value_rate", dto.getCurUnit(), json)
+                                .whenComplete((result, e) -> {
+                                    if (e != null) {
+                                        log.error(e.getMessage());
+                                    } else {
+                                        log.info(result.toString());
+                                    }
+                                });
+                    }
                 }
             }
